requirements.txt
django==1.8.6
djangorestframework==3.2.5
Markdown==2.6.4
psycopg2==2.6.1
